backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
beautifulsoup4
jinja2
python-dotenv
requests
